Fixing foreign keys errors and logs model events

+++ database/migrations/2018_06_09_151005_create_invoice_shift_table.php
@@ -15,7 +15,7 @@ class CreateInvoiceShiftTable extends Migration
     {
         Schema::create('invoice_shift', function (Blueprint $table) {
             $table->bigInteger('invoice_id')->unsigned();
-            $table->bigInteger('shift_id')->unsigned();
+            $table->integer('shift_id')->unsigned();
 
             $table->foreign('invoice_id')->references('id')->on('invoices')
             ->onUpdate('cascade')->onDelete('cascade');
