+++ database/migrations/2018_06_09_151155_create_product_room_table.php
@@ -14,8 +14,8 @@ class CreateProductRoomTable extends Migration
     public function up()
     {
         Schema::create('product_room', function (Blueprint $table) {
-            $table->bigInteger('product_id')->unsigned();
-            $table->bigInteger('room_id')->unsigned();
+            $table->integer('product_id')->unsigned();
+            $table->integer('room_id')->unsigned();
             $table->integer('quantity');
 
             $table->foreign('product_id')->references('id')->on('products')
